fix(security): clamp rate-limit counts in int64 before narrowing (CodeQL #39, #40)
ratelimit.Remaining takes the used count as int64 and clamps it to [0, limit] before any conversion, so a Lua reply outside int range can never wrap.

+++ ratelimit/memory/limiter.go
@@ -147,7 +147,7 @@ func (l *Limiter) AllowNamedResult(bucket, key string) (ratelimit.Result, error)
 			RetryAfter: retryAfter,
 			Reason:     retryReason,
 			Limit:      lim.Limit,
-			Remaining:  ratelimit.Remaining(lim.Limit, len(ts)),
+			Remaining:  ratelimit.Remaining(lim.Limit, int64(len(ts))),
 			Window:     lim.Window,
 			Cooldown:   lim.Cooldown,
 		}, nil
@@ -162,7 +162,7 @@ func (l *Limiter) AllowNamedResult(bucket, key string) (ratelimit.Result, error)
 	return ratelimit.Result{
 		Allowed:   true,
 		Limit:     lim.Limit,
-		Remaining: ratelimit.Remaining(lim.Limit, len(ts)),
+		Remaining: ratelimit.Remaining(lim.Limit, int64(len(ts))),
 		Window:    lim.Window,
 		Cooldown:  lim.Cooldown,
 	}, nil
